fix(Firmware): resolve all -Wall -Wextra -Werror warnings - easing.h: change easing_frac_t to uint16_t (fix Q1.15 overflow) - SPI.h: suppress unused variable in volatile read macro - gpio.c, timer.c, wdg.c: fix sign-compare (int vs size_t) - rtc.c: suppress unused parameters in stub function - DeviceObject.h: suppress unused params in virtual defaults - HAL.cpp: remove unused lambda parameter name - DP_Alarm.cpp: fix sign-compare, replace memset with value-init - DP_Ctrl.cpp, DP_Clock.cpp: fix loop variable signedness - MotorCtrl.cpp: fix sign-compare in array bounds checks - CMakeLists.txt: enable -Werror without any -Wno-* suppressions

+++ Firmware/App/Service/DataProc/DP_Ctrl.cpp
@@ -68,7 +68,7 @@ DP_Ctrl::DP_Ctrl(DataNode* node)
         return;
     }
 
-    for (int i = 0; i < CM_ARRAY_SIZE(_mctrl); i++) {
+    for (size_t i = 0; i < CM_ARRAY_SIZE(_mctrl); i++) {
         _mctrl[i].setID(i);
         _mctrl[i].setNode(node);
         _mctrl[i].setDevice(_devMotor);
@@ -168,7 +168,7 @@ int DP_Ctrl::onNotify(const Ctrl_Info_t* info)
 void DP_Ctrl::onTimer()
 {
     int activeMotorCount = 0;
-    for (int i = 0; i < CM_ARRAY_SIZE(_mctrl); i++) {
+    for (size_t i = 0; i < CM_ARRAY_SIZE(_mctrl); i++) {
         activeMotorCount += _mctrl[i].timerHandler();
     }
